# api/cmd/api/docs/swagger.yaml
definitions:
  github_com_Caqil_megapdf-api_internal_models.Account:
    properties:
      accessToken:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expiresAt:
        type: integer
      id:
        type: integer
      idToken:
        type: string
      provider:
        type: string
      providerAccountId:
        type: string
      refreshToken:
        type: string
      scope:
        type: string
      sessionState:
        type: string
      tokenType:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.ApiKey:
    properties:
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      key:
        type: string
      lastUsed:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.PasswordResetToken:
    properties:
      createdAt:
        type: string
      email:
        type: string
      expires:
        type: string
      id:
        type: string
      token:
        type: string
    type: object
...
  github_com_Caqil_megapdf-api_internal_models.Session:
    properties:
      expires:
        type: string
      id:
        type: string
      sessionToken:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
...
  github_com_Caqil_megapdf-api_internal_models.Transaction:
    properties:
      amount:
        type: number
      balanceAfter:
        type: number
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      paymentId:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.UsageStats:
    properties:
      count:
        type: integer
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      operation:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.User:
    properties:
      accounts:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
        type: array
      apiKeys:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
        type: array
      balance:
        type: number
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: string
      freeOperationsReset:
        type: string
      freeOperationsUsed:
        type: integer
      id:
        type: string
      image:
        type: string
      isEmailVerified:
        type: boolean
      name:
        type: string
      password:
        type: string
      role:
        type: string
      sessions:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
        type: array
      transactions:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
        type: array
      updatedAt:
        type: string
      usageStats:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
        type: array
      verificationToken:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Creates a new account with the provided details
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Retrieves an account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an account
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Updates an account with the provided details
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an account
      tags:
      - accounts
  /api-keys:
    post:
      consumes:
      - application/json
      description: Creates a new API key with the provided details
      parameters:
      - description: ApiKey data
        in: body
        name: apiKey
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new API key
      tags:
      - api-keys
  /api-keys/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an API key by its ID
      parameters:
      - description: ApiKey ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an API key
      tags:
      - api-keys
    get:
      consumes:
      - application/json
      description: Retrieves an API key by its ID
      parameters:
      - description: ApiKey ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an API key
      tags:
      - api-keys
    put:
      consumes:
      - application/json
      description: Updates an API key with the provided details
      parameters:
      - description: ApiKey ID
        in: path
        name: id
        required: true
        type: string
      - description: ApiKey data
        in: body
        name: apiKey
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an API key
      tags:
      - api-keys
  /password-reset-tokens:
    post:
      consumes:
      - application/json
      description: Creates a new password reset token with the provided details
      parameters:
      - description: PasswordResetToken data
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PasswordResetToken'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PasswordResetToken'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new password reset token
      tags:
      - password-reset-tokens
  /password-reset-tokens/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a password reset token by its ID
      parameters:
      - description: PasswordResetToken ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a password reset token
      tags:
      - password-reset-tokens
    get:
      consumes:
      - application/json
      description: Retrieves a password reset token by its ID
      parameters:
      - description: PasswordResetToken ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.PasswordResetToken'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a password reset token
      tags:
      - password-reset-tokens
...
swagger: "2.0"